Remove overly zealous error check in atsas-dat parser (fixes #28)
`saxs_document_add_property` can return NULL either because of a
memory allocation error or when one of the input strings has zero
length. The atsas-dat parser simply always tries to create a
property even when none was found, relying on the fact that
zero-length properties will not be created.

However, this means that a NULL return value does not necessarily
mean a memory allocation error. Therefore this patch removes the
checks on the return value.